+++ HISTORY.rst
@@ -16,6 +16,7 @@
 **Bug fixes**
 
 - 872_: [Linux] can now compile on Linux by using MUSL C library.
+- 985_: [Windows] Fix a crash in `Process.open_files` when the worker thread for `NtQueryObject` times out.
 
 5.1.3
 =====